Update README.md with an audio patching guide
One of macOS's most imfamous post-install issues is a glitch with the AUX port. Everytime I update the EFI always causes a problem with my audio patch. "Give one a fish you feed them for a day but teach one how to fish and you feed them for a lifetime."  So, here is a guide on how to fix the audio yourself.